% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.SNPDataLong}
\alias{print.summary.SNPDataLong}
\title{Print method for SNPDataLong summary}
\usage{
\method{print}{summary.SNPDataLong}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.SNPDataLong}.}

\item{...}{Further arguments (currently unused).}
}
\value{
The input \code{x}, returned invisibly.
}
\description{
Displays the contents of a \code{summary.SNPDataLong} object on the console.
}
