% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPkit-package.R
\docType{package}
\name{SNPkit}
\alias{SNPkit-package}
\alias{SNPkit}
\title{SNPkit: S4 tools for reading and organizing genetic data}
\description{
Utilities for reading, cleaning, summarizing, and modeling SNP genotype data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://viniciusjunqueira.github.io/SNPkit/}
  \item \url{https://github.com/viniciusjunqueira/SNPkit}
  \item Report bugs at \url{https://github.com/viniciusjunqueira/SNPkit/issues}
}

}
\author{
Vinícius Junqueira \email{junqueiravinicius@hotmail.com}
Roberto Higa \email{roberto.higa@embrapa.br}
Fernando Flores Cardoso \email{fernando.cardoso@embrapa.br}
Marcos Jun Iti Yokoo \email{marcos.yokoo@embrapa.br}
}
\keyword{internal}
