% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{doPCA}
\alias{doPCA}
\title{Do genome relationship matrix PCA}
\usage{
doPCA(genotypes)
}
\arguments{
\item{genotypes}{Genotype matrix.}
}
\value{
List containing `pcs` (principal components) and `eigen` (eigenvalues).
}
\description{
Performs PCA using the genome relationship matrix (GRM).
}
\examples{
\donttest{
set.seed(1)
mat <- matrix(sample(as.raw(1:3), 200, TRUE), nrow = 10, ncol = 20)
geno <- methods::new("SnpMatrix", mat)
rownames(geno) <- paste0("S", 1:10)
colnames(geno) <- paste0("SNP", 1:20)
res <- doPCA(geno)
str(res)
}

}
